only include libloading when using `std`
When we do not use `std`, we cannot load things dynamically with
`libloading`. By its very nature, `libloading` will probably always
include a lot of platform-specific code, making it not depend on `std`
unrealistic. So, when someone says they do not want to use `std`, we
should not use `libloading`. As long as downstream does not try to
dynamically load something with `default-features=false`, this should
never be a problem.

What happens if downstream does try that, however? Well, then the Rust
compiler with print an error saying they are trying to use a crate that
does not exist. Because we default to using `std` (as is the standard
for most optionally `no_std` crates), this should not cause any
breakage.